refactor: move app.node.module.ts and app.browser.module.ts to app/
