+++ src/app/app.browser.module.ts
@@ -3,10 +3,10 @@ import { FormsModule } from '@angular/forms';
 import { RouterModule } from '@angular/router';
 import { UniversalModule, isBrowser, isNode } from 'angular2-universal/browser'; // for AoT we need to manually split universal packages
 
-import { AppComponent } from './app/app.component';
-import { HomeComponent } from './app/home.component';
-import { routes } from './app/app.routes';
-import { Cache } from './app/universal-cache';
+import { AppComponent } from './app.component';
+import { HomeComponent } from './home.component';
+import { routes } from './app.routes';
+import { Cache } from './universal-cache';
 
 @NgModule({
   bootstrap: [ AppComponent ],
