+++ src/server.ts
@@ -14,7 +14,7 @@ import { enableProdMode } from '@angular/core';
 import { createEngine } from 'angular2-express-engine';
 
 // App
-import { MainModule } from './app.node.module';
+import { MainModule } from './app/app.node.module';
 
 // enable prod for faster renders
 enableProdMode();
